requirements.txt
flask
Flask_Session
requests
psycopg2-binary
# psycopg2
python-dotenv